chore(sink): separated otel metrics into config encoder and service files
